+++ tests/constants.py
@@ -1,0 +1,11 @@
+TOKENIZER_ROUND_TRIP_STRINGS = [
+    "",
+    " ",
+    "hello",
+    " hello",
+    "two words",
+    " two words",
+    " two words ",
+    "two words ",
+    "’",
+]
