+++ tests/unit/test_tokenizers.py
@@ -3,6 +3,8 @@ import tiktoken
 
 from guidance import models
 
+from tests.constants import TOKENIZER_ROUND_TRIP_STRINGS
+
 # These are not _strictly_ unit tests, since they
 # refer to specific tokenisers. However, tokenisers
 # are small, so if the tokeniser can be loaded
@@ -12,18 +14,6 @@ from guidance import models
 
 # There is a separate file for the llamacpp models
 # since the tokenisers cannot be loaded separately
-
-ROUND_TRIP_STRINGS = [
-    "",
-    " ",
-    "hello",
-    " hello",
-    "two words",
-    " two words",
-    " two words ",
-    "two words ",
-    "’",
-]
 
 
 class TestTransformerTokenizers:
@@ -42,7 +32,7 @@ class TestTransformerTokenizers:
         assert my_tok is not None
 
     @pytest.mark.parametrize("model_name", TRANSFORMER_MODELS)
-    @pytest.mark.parametrize("target_string", ROUND_TRIP_STRINGS)
+    @pytest.mark.parametrize("target_string", TOKENIZER_ROUND_TRIP_STRINGS)
     def test_string_roundtrip(self, model_name: str, target_string: str):
         my_tok = models.TransformersTokenizer(model=model_name, transformers_tokenizer=None)
 
@@ -57,7 +47,7 @@ class TestTiktoken:
     MODELS = ["gpt-3.5-turbo", "gpt-4"]
 
     @pytest.mark.parametrize("model_name", MODELS)
-    @pytest.mark.parametrize("target_string", ROUND_TRIP_STRINGS)
+    @pytest.mark.parametrize("target_string", TOKENIZER_ROUND_TRIP_STRINGS)
     def test_string_roundtrip(self, model_name: str, target_string: str):
         my_tik = tiktoken.encoding_for_model(model_name)
         my_tok = models._grammarless.GrammarlessTokenizer(my_tik)
